feat: add flag for saving the proxy address explicitly on the machine
Default behavior has been changed to NOT set the proxy explicitly, and will only rely on the PAC in-memory file.

Utilizing only the PAC file is useful for most scenarios and it comes with the benefit of falling back to a direct connection should the proxy connection fail.

Applying the proxy explicitly on the machine will allow a wider range of applications that don't seem to work with just a PAC file.